Simplify callbacks in Renderer interface
The callbacks used to return bools, but none of the actual
implementations return false, always true. So in order to make further
refactorings simpler, make the interface reflect the inner workings: no
more return values, no more conditionals.

Note that this probably puts an end to the idea of an interface
compatibility with v1.x. This is a semantic change, so even though the
interface can be preserved, the new internal functionality will no
longer match the old implementations.