Add another method to take over writes to output
We need two different ways: one that passes the output through, but
grabs a copy and another that redirects the output completely, buffers
it, but doesn't write to the destination.

All the places where captureWrites is used need revisiting, but this
commit contains the first use case for it: BeginHeader/EndHeader.