Implement footnotes
Parse footnotes and insert their list at the end of the tree.

This commit does some changes to the test suite. The changes are only to
the newlines. Turns out the former implementation of footnotes did
slightly different spacing when rendering the footnotes list. The new
implementation produces the list that is compatible with the rest of the
package.